english_talker_project/requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
aiohttp==3.9.1
python-multipart==0.0.6
edge-tts==6.1.9
pyttsx3==2.90
gtts==2.4.0
jinja2==3.1.2
python-dotenv==1.0.0
discord-webhook==1.3.0
ffmpeg-python==0.2.0
langchain==0.0.289